move modal's logic to context component
